refactor provider layers
the outer package should be responsible for components, the inner
package should be responsible for infrastructure
